+++ packages/ricos-editor/web/src/RicosEditor.spec.tsx
@@ -70,7 +70,7 @@ describe('RicosEditor', () => {
     expect(rceProps.config).toHaveProperty('wix-draft-plugin-hashtag');
   });
   it('should render with themeStrategy output', () => {
-    const rceProps = getRCE({ theme: createTheme() }).props();
+    const rceProps = getRCE({ theme: createTheme({}) }).props();
     expect(rceProps).toHaveProperty('theme');
     expect(rceProps.theme).toHaveProperty('modalTheme');
   });