+++ packages/ricos-theme/web/src/ThemeGenerator.ts
@@ -1,33 +1,32 @@
 import { PaletteColors } from 'wix-rich-content-common';
-import * as utils from './themes/utils';
-import { palettes, assertPalette, COLORS } from './palettes';
-import getEditorCommonTheme from './themes/editor-common';
-import getEditorTheme from './themes/editor';
-import getViewerTheme from './themes/viewer';
-import getCommonStyles from './themes/common';
-import { merge } from 'lodash';
-import { PalettePreset, Palette, Color, ThemeGeneratorFunction } from 'ricos-common';
+import * as utils from './themeUtils';
+import { presets, assertWixPalette, COLORS, isRicosPalette, getColorValue } from './palettes';
+import { ThemeGeneratorFunction, RicosTheme } from 'ricos-common';
 
-/* eslint-disable camelcase */
-
-const PALETTE_PRESETS: { [propName in PalettePreset]: Palette } = { darkTheme: palettes.darkTheme };
-
-const getColorByCode = (palette: Palette, code: number): Color => {
-  const idx = code <= 5 ? code - 1 : code - 6;
-  return palette[idx];
+const createCssVars = (colors: PaletteColors) => {
+  const { adaptForeground, toRgbTuple } = utils;
+  const { textColor, bgColor, actionColor } = colors;
+  return `
+  * {
+    --ricos-text-color: ${textColor};
+    --ricos-text-color-tuple: ${toRgbTuple(textColor)};
+    --ricos-action-color: ${actionColor};
+    --ricos-action-color-tuple: ${toRgbTuple(actionColor)};
+    --ricos-action-color-fallback: ${adaptForeground(actionColor)};
+    --ricos-action-color-fallback-tuple: ${toRgbTuple(adaptForeground(actionColor))};
+    --ricos-background-color: ${bgColor};
+    --ricos-background-color-tuple: ${toRgbTuple(bgColor)};
+  }\n`;
 };
-
-const getColorValue = (palette: Palette, code: number): string =>
-  getColorByCode(palette, code).value;
 
 export default class ThemeGenerator {
   isViewer: boolean;
   themeGeneratorFunctions: ThemeGeneratorFunction[];
-  palette?: Palette;
+  palette?: PaletteColors;
 
   constructor(
     isViewer: boolean,
-    palette?: Palette | PalettePreset,
+    palette?: RicosTheme['palette'],
     themeGeneratorFunctions: ThemeGeneratorFunction[] = []
   ) {
     this.setPalette(palette);
@@ -35,38 +34,32 @@ export default class ThemeGenerator {
     this.isViewer = isViewer;
   }
 
-  setPalette(palette?: string | Palette) {
+  setPalette(palette?: RicosTheme['palette']) {
+    if (!palette) return;
     if (typeof palette === 'string') {
-      if (palette in PALETTE_PRESETS) {
-        this.palette = PALETTE_PRESETS[palette];
+      if (palette in presets) {
+        this.palette = presets[palette];
       } else {
-        throw Error(
-          `Palette ${palette} is unknown. Supported themes: ${PALETTE_PRESETS.toString()}`
-        );
+        throw Error(`Palette ${palette} is unknown. Supported themes: ${presets.toString()}`);
       }
-    } else {
-      assertPalette(palette);
+    } else if (Array.isArray(palette)) {
+      assertWixPalette(palette);
+      this.palette = {
+        actionColor: getColorValue(palette, COLORS.ACTION_COLOR),
+        bgColor: getColorValue(palette, COLORS.BG_COLOR),
+        textColor: getColorValue(palette, COLORS.TEXT_COLOR),
+      };
+    } else if (isRicosPalette(palette)) {
       this.palette = palette;
     }
   }
 
-  getStylesObject() {
+  getStylesString() {
     if (!this.palette) {
-      return {};
+      return '';
     }
-    const colors: PaletteColors = {
-      actionColor: getColorValue(this.palette, COLORS.ACTION_COLOR),
-      bgColor: getColorValue(this.palette, COLORS.BG_COLOR),
-      textColor: getColorValue(this.palette, COLORS.TEXT_COLOR),
-    };
-
-    const pluginThemes = this.themeGeneratorFunctions.map(themeGen => themeGen(colors, utils));
-    const appStyles = !this.isViewer
-      ? merge(getEditorCommonTheme(colors), getEditorTheme(colors, utils))
-      : getViewerTheme(colors, utils);
-
-    return merge(getCommonStyles(colors), appStyles, ...pluginThemes);
+    const colors = this.palette;
+    this.themeGeneratorFunctions.forEach(themeGen => themeGen(colors, utils));
+    return createCssVars(colors);
   }
 }
-
-export { PALETTE_PRESETS };